services/investment-service/requirements.txt
# requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
pydantic>=2.0.0
psycopg2-binary>=2.9.1
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.5
python-dotenv>=0.19.0

# Development dependencies
-r requirements-dev.txt